extract network, rename txnbuild->transaction
