+++ exp/txnbuild/transaction.go
@@ -12,20 +12,6 @@ import (
 	"github.com/stellar/go/xdr"
 )
 
-// StellarNetwork is a global setting that sets the choice of network to use.
-var StellarNetwork = network.TestNetworkPassphrase
-
-// UseTestNetwork sets the global network setting to use the Stellar TestNet.
-func UseTestNetwork() {
-	StellarNetwork = network.TestNetworkPassphrase
-}
-
-// UseTestNetwork sets the global network setting to use the Stellar Public network
-// (i.e. production).
-func UsePublicNetwork() {
-	StellarNetwork = network.PublicNetworkPassphrase
-}
-
 // TODO: Replace use of Horizon Account with simpler Account object here
 // type Account struct {
 // 	AccountID string
